+++ Android/app/src/main/java/com/khanabook/lite/pos/data/remote/dto/PushSyncResponse.kt
@@ -2,13 +2,10 @@ package com.khanabook.lite.pos.data.remote.dto
 
 import com.google.gson.annotations.SerializedName
 
-/**
- * Data class matching the server's PushSyncResponse shape.
- * Used to handle sync responses without crashing.
- */
+
 data class PushSyncResponse(
     @SerializedName("successfulLocalIds")
-    val successfulLocalIds: List<Int>,
+    val successfulLocalIds: List<Long>,
     @SerializedName("failedLocalIds")
-    val failedLocalIds: List<Int>
+    val failedLocalIds: List<Long>
 )
